# docs/api.yaml
openapi: 3.1.0
info:
  title: WeFixIt HTTP API
  version: 1.0.0
  description: Session-authenticated endpoints used by the server-rendered application.
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: wefixit_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
paths:
...
  /payment/checkout:
    post:
      summary: Create a Stripe Checkout session for confirmed work
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses:
        '200': { description: Checkout URL returned }
        '409': { description: Work, quote, or confirmation is incomplete }
  /payment/receipt/{bookingId}/pdf:
    get:
      summary: Download a paid booking receipt as a PDF
      security: [{ sessionCookie: [] }]
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Receipt PDF
          content:
            application/pdf: {}
        '404': { description: Paid receipt not found }